MCP/mcp-sqlite-agent-implementation/requirements.txt
requests     
python-dotenv
mcp[cli]       
openai